+++ src/algorithms/games/mod.rs
@@ -6,3 +6,4 @@
 
 pub mod barabasi;
 pub mod erdos_renyi;
+pub mod growing_random;